+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 This document contains the Spec2nii release history in reverse chronological order.
 
+0.7.0 (Saturday 5th August 2023)
+--------------------------------
+- Fixed a bug in Philips Classic DICOM orientations (supplementing the fixes to Enhanced DICOM in `0.6.11`)
+- Updated Hyper sequence special case for Siemens twix and GE.
+
+0.6.11 (Friday 28th July 2023)
+------------------------------
+- Fixed a bug in Philips spar/sdat orientations - this will affect voxels with rotations in more than one axes.
+- Fixed a bug in Philips DICOM orientations - this will affect voxels with rotations in more than one axes.
+- Python 3.7 now in [end of life](https://devguide.python.org/versions/) status and is no longer supported.
+- `spec2nii dump/extract/insert` can now be used to inspect and fix non-compliant NIfTI-MRS files.
+
 0.6.10 (Monday 10th July 2023)
 ------------------------------
-- Stop tests being vendered as a top-level package. Bug introduced in `0.6.9`.
+- Stop tests being vendored as a top-level package. Bug introduced in `0.6.9`.
 
 0.6.9 (Friday 7th July 2023)
 ----------------------------